apply reducer to screen
