+++ navigation/MainTabNavigator.js
@@ -11,7 +11,7 @@ import MyPageScreen from "../screens/MyPage"
 import PlanedListScreen from '../screens/PlannedList/index'
 import MainScreen from '../screens/Main';
 import NoteDetailScreen from '../screens/NoteDetail/index';
-import NoteEditScreen from '../screens/NoteEdit';
+import NoteEditScreen from '../screens/NoteEdit/index';
 import NoteListScreen from '../screens/NoteList/index';
 import SignScreen from '../screens/Sign';
